Add isValid to POST /confirm-email
Make sure a request with an invalid access code still return a successful
response. It makes monitoring the server easier, because it's an expect
behavior that a user enters an incorrect access code. Http error should
preferably only be used for unexpected behaviors.

Adding `isValid` to the response, indicating if the accessCode is valid.